Fix Mercprofile usStrategicInsertionData for bigmaps
NPCs would not appear in the map if their gridno was over 65535 in large maps

+++ Tactical/soldier profile type.h
@@ -934,7 +934,7 @@ public:
 	UINT8 ubLastDateSpokenTo;
 	UINT8 bLastQuoteSaidWasSpecial;
 	INT8	bSectorZ;
-	UINT16 usStrategicInsertionData;
+	UINT32 usStrategicInsertionData;
 	INT8 bFriendlyOrDirectDefaultResponseUsedRecently;
 	INT8 bRecruitDefaultResponseUsedRecently;
 	INT8 bThreatenDefaultResponseUsedRecently;
